+++ controlador/integrais/integralGaussHermite.h
@@ -5,10 +5,10 @@
 #include <cmath>
 #include "../parser/parser.h"
 
-float gaussHermite (const int &qtdPontos, expre::expre *equacao);
+double gaussHermite (const int &qtdPontos, expre::expre *equacao);
 
-float gaussHermiteN2 (expre::expre *equacao);
-float gaussHermiteN3 (expre::expre *equacao);
-float gaussHermiteN4 (expre::expre *equacao);
+double gaussHermiteN2 (expre::expre *equacao);
+double gaussHermiteN3 (expre::expre *equacao);
+double gaussHermiteN4 (expre::expre *equacao);
 
 #endif //INTEGRAL_GAUSS_HERMITE_H
