+++ controlador/parser/parser.cpp
@@ -97,7 +97,7 @@
 ** Parâmetros: A string e a primeira posição do número na string
 ** Retornos: O número lido
 */
-float lerConstante (std::string &expressao, int &percorrer) {
+double lerConstante (std::string &expressao, int &percorrer) {
   int quant = 0, inicio = percorrer, tam = expressao.size();
         
   while(((percorrer < tam) && ((expressao[percorrer] >= '0') && (expressao[percorrer] <= '9')))
